+++ ReleaseNotes.md
@@ -1,5 +1,13 @@
 ## Release Notes ##
 
+**Version 0.4.0 (compatible with SurrealDB version 1.4.0):**
+- code attribute of a result is now containing http status code or error code for websocket
+- SHOW CHANGES uses current date-time if not specified, because SHOW does not work without SINCE
+- fix live queries url links
+- live query for QL is now can use VALUE and custom query
+- DEFINE TABLE now can use TYPE and INCLUDE ORIGINAL
+- minor improvements for documentation, examples and tests
+
 **Version 0.3.1 (compatible with SurrealDB version 1.3.1):**
 - increment default timeout to 15 seconds
 - minor improvements for documentation